Merge v10 work into core

+++ concrete/controllers/dialog/logs/delete_all.php
@@ -26,26 +26,28 @@ class DeleteAll extends BackendInterfaceController
 
     public function submit()
     {
-        /** @var ResponseFactory $responseFactory */
-        $responseFactory = $this->app->make(ResponseFactory::class);
-        /** @var EditResponse $editResponse */
-        $editResponse = new EditResponse();
+        if ($this->validateAction()) {
+            /** @var ResponseFactory $responseFactory */
+            $responseFactory = $this->app->make(ResponseFactory::class);
+            /** @var EditResponse $editResponse */
+            $editResponse = new EditResponse();
 
-        if ($this->canAccess()) {
-            /** @var Connection $db */
-            $db = $this->app->make(Connection::class);
+            if ($this->canAccess()) {
+                /** @var Connection $db */
+                $db = $this->app->make(Connection::class);
 
-            /** @noinspection PhpUnhandledExceptionInspection */
-            /** @noinspection SqlDialectInspection */
-            /** @noinspection SqlNoDataSourceInspection */
-            $db->executeQuery("TRUNCATE TABLE Logs");
+                /** @noinspection PhpUnhandledExceptionInspection */
+                /** @noinspection SqlDialectInspection */
+                /** @noinspection SqlNoDataSourceInspection */
+                $db->executeQuery("TRUNCATE TABLE Logs");
 
-            $this->flash('success', t('Log cleared successfully.'));
+                $this->flash('success', t('Log cleared successfully.'));
             $editResponse->setRedirectURL((string) $this->app->make(ResolverManagerInterface::class)->resolve(['/dashboard/reports/logs']));
-        } else {
-            $editResponse->setMessage(t('Access denied'));
-        }
+            } else {
+                $editResponse->setMessage(t('Access denied'));
+            }
 
-        return $responseFactory->json($editResponse->getJSONObject());
+            return $responseFactory->json($editResponse->getJSONObject());
+        }
     }
 }
